refactor: Update SRP helpers
