+++ packages/auth/src/providers/cognito/utils/srp/BigInteger/index.ts
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import BigInteger from './BigInteger';
-
+export { AuthBigInteger } from './types';
 export { BigInteger };
