+++ packages/auth/src/providers/cognito/utils/srp/calculate/index.ts
@@ -1,0 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { calculateA } from './calculateA';
+export { calculateS } from './calculateS';
+export { calculateU } from './calculateU';
