+++ packages/auth/src/providers/cognito/utils/srp/index.ts
@@ -1,0 +1,7 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { getAuthenticationHelper } from './getAuthenticationHelper';
+export { getBytesFromHex } from './getBytesFromHex';
+export { getNowString } from './getNowString';
+export { getSignatureString } from './getSignatureString';
